+++ src/DMLQueryBuilder.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Oracle;
 
+use Generator;
 use InvalidArgumentException;
 use JsonException;
 use Yiisoft\Db\Constraint\Constraint;
@@ -11,19 +12,90 @@ use Yiisoft\Db\Exception\Exception;
 use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
 use Yiisoft\Db\Expression\Expression;
+use Yiisoft\Db\Expression\ExpressionInterface;
 use Yiisoft\Db\Query\DMLQueryBuilder as AbstractDMLQueryBuilder;
+use Yiisoft\Db\Query\Query;
 use Yiisoft\Db\Query\QueryBuilderInterface;
 use Yiisoft\Db\Query\QueryInterface;
+use Yiisoft\Strings\NumericHelper;
 
 use function implode;
 use function ltrim;
 use function strrpos;
+use function count;
+use function is_string;
+use function reset;
 
 final class DMLQueryBuilder extends AbstractDMLQueryBuilder
 {
     public function __construct(private QueryBuilderInterface $queryBuilder)
     {
         parent::__construct($queryBuilder);
+    }
+
+    /**
+     * @psalm-suppress MixedArrayOffset
+     */
+    public function batchInsert(string $table, array $columns, iterable|Generator $rows, array &$params = []): string
+    {
+        if (empty($rows)) {
+            return '';
+        }
+
+        $schema = $this->schema;
+
+        if (($tableSchema = $schema->getTableSchema($table)) !== null) {
+            $columnSchemas = $tableSchema->getColumns();
+        } else {
+            $columnSchemas = [];
+        }
+
+        $values = [];
+
+        /** @psalm-var string[][] $rows */
+        foreach ($rows as $row) {
+            $vs = [];
+            foreach ($row as $i => $value) {
+                if (isset($columns[$i], $columnSchemas[$columns[$i]])) {
+                    /** @var mixed $value */
+                    $value = $columnSchemas[$columns[$i]]->dbTypecast($value);
+                }
+
+                if (is_string($value)) {
+                    /** @var mixed $value */
+                    $value = $this->quoter->quoteValue($value);
+                } elseif (is_float($value)) {
+                    /* ensure type cast always has . as decimal separator in all locales */
+                    $value = NumericHelper::normalize($value);
+                } elseif ($value === false) {
+                    $value = 0;
+                } elseif ($value === null) {
+                    $value = 'NULL';
+                } elseif ($value instanceof ExpressionInterface) {
+                    $value = $this->queryBuilder->buildExpression($value, $params);
+                }
+
+                /** @var mixed */
+                $vs[] = $value;
+            }
+
+            /** @psalm-var string[] $vs */
+            $values[] = '(' . implode(', ', $vs) . ')';
+        }
+
+        if (empty($values)) {
+            return '';
+        }
+
+        /** @psalm-var string[] $columns */
+        foreach ($columns as $i => $name) {
+            $columns[$i] = $this->quoter->quoteColumnName($name);
+        }
+
+        $tableAndColumns = ' INTO ' . $this->quoter->quoteTableName($table)
+            . ' (' . implode(', ', $columns) . ') VALUES ';
+
+        return 'INSERT ALL ' . $tableAndColumns . implode($tableAndColumns, $values) . ' SELECT 1 FROM SYS.DUAL';
     }
 
     /**
@@ -47,8 +119,7 @@ final class DMLQueryBuilder extends AbstractDMLQueryBuilder
         $usingValues = null;
         $constraints = [];
 
-        /** @var Constraint[] $constraints */
-        [$uniqueNames, $insertNames, $updateNames] = $this->queryBuilder->prepareUpsertColumns(
+        [$uniqueNames, $insertNames, $updateNames] = $this->prepareUpsertColumns(
             $table,
             $insertColumns,
             $updateColumns,
@@ -65,14 +136,14 @@ final class DMLQueryBuilder extends AbstractDMLQueryBuilder
         }
 
         $onCondition = ['or'];
-        $quotedTableName = $this->queryBuilder->quoter()->quoteTableName($table);
+        $quotedTableName = $this->quoter->quoteTableName($table);
 
         foreach ($constraints as $constraint) {
             $columnNames = $constraint->getColumnNames() ?? [];
             $constraintCondition = ['and'];
             /** @psalm-var string[] $columnNames */
             foreach ($columnNames as $name) {
-                $quotedName = $this->queryBuilder->quoter()->quoteColumnName($name);
+                $quotedName = $this->quoter->quoteColumnName($name);
                 $constraintCondition[] = "$quotedTableName.$quotedName=\"EXCLUDED\".$quotedName";
             }
 
@@ -81,7 +152,7 @@ final class DMLQueryBuilder extends AbstractDMLQueryBuilder
 
         $on = $this->queryBuilder->buildCondition($onCondition, $params);
         /** @psalm-var string[] $placeholders */
-        [, $placeholders, $values, $params] = $this->queryBuilder->prepareInsertValues($table, $insertColumns, $params);
+        [, $placeholders, $values, $params] = $this->prepareInsertValues($table, $insertColumns, $params);
 
         if (!empty($placeholders)) {
             $usingSelectValues = [];
@@ -96,7 +167,7 @@ final class DMLQueryBuilder extends AbstractDMLQueryBuilder
 
         $insertValues = [];
         $mergeSql = 'MERGE INTO '
-            . $this->queryBuilder->quoter()->quoteTableName($table)
+            . $this->quoter->quoteTableName($table)
             . ' '
             . 'USING (' . ($usingValues ?? ltrim((string) $values, ' '))
             . ') "EXCLUDED" '
@@ -104,7 +175,7 @@ final class DMLQueryBuilder extends AbstractDMLQueryBuilder
 
         /** @psalm-var string[] $insertNames */
         foreach ($insertNames as $name) {
-            $quotedName = $this->queryBuilder->quoter()->quoteColumnName($name);
+            $quotedName = $this->quoter->quoteColumnName($name);
 
             if (strrpos($quotedName, '.') === false) {
                 $quotedName = '"EXCLUDED".' . $quotedName;
@@ -123,7 +194,7 @@ final class DMLQueryBuilder extends AbstractDMLQueryBuilder
             $updateColumns = [];
             /** @psalm-var string[] $updateNames */
             foreach ($updateNames as $name) {
-                $quotedName = $this->queryBuilder->quoter()->quoteColumnName($name);
+                $quotedName = $this->quoter->quoteColumnName($name);
 
                 if (strrpos($quotedName, '.') === false) {
                     $quotedName = '"EXCLUDED".' . $quotedName;
@@ -133,9 +204,67 @@ final class DMLQueryBuilder extends AbstractDMLQueryBuilder
         }
 
         /** @psalm-var string[] $updates */
-        [$updates, $params] = $this->queryBuilder->prepareUpdateSets($table, $updateColumns, (array) $params);
+        [$updates, $params] = $this->prepareUpdateSets($table, $updateColumns, (array) $params);
         $updateSql = 'UPDATE SET ' . implode(', ', $updates);
 
         return "$mergeSql WHEN MATCHED THEN $updateSql WHEN NOT MATCHED THEN $insertSql";
+    }
+
+    protected function prepareInsertValues(string $table, array|QueryInterface $columns, array $params = []): array
+    {
+        /**
+         * @var array $names
+         * @var array $placeholders
+         */
+        [$names, $placeholders, $values, $params] = parent::prepareInsertValues($table, $columns, $params);
+
+        if (!$columns instanceof Query && empty($names)) {
+            $tableSchema = $this->schema->getTableSchema($table);
+
+            if ($tableSchema !== null) {
+                $tableColumns = $tableSchema->getColumns();
+                $columns = !empty($tableSchema->getPrimaryKey())
+                    ? $tableSchema->getPrimaryKey() : [reset($tableColumns)->getName()];
+                foreach ($columns as $name) {
+                    /** @var mixed */
+                    $names[] = $this->quoter->quoteColumnName($name);
+                    $placeholders[] = 'DEFAULT';
+                }
+            }
+        }
+
+        return [$names, $placeholders, $values, $params];
+    }
+
+    public function resetSequence(string $tableName, array|int|string|null $value = null): string
+    {
+        $tableSchema = $this->schema->getTableSchema($tableName);
+
+        if ($tableSchema === null) {
+            throw new \Yiisoft\Db\Exception\InvalidArgumentException("Unknown table: $tableName");
+        }
+
+        $sequenceName = $tableSchema->getSequenceName();
+        if ($sequenceName === null) {
+            throw new InvalidArgumentException("There is no sequence associated with table: $tableName");
+        }
+
+        if ($value !== null) {
+            $value = (int) $value;
+        } elseif (count($tableSchema->getPrimaryKey()) > 1) {
+            throw new InvalidArgumentException("Can't reset sequence for composite primary key in table: $tableName");
+        }
+
+        /**
+         *  Oracle needs at least many queries to reset sequence (see adding transactions and/or use alter method to avoid grants issue?)
+         */
+        return 'declare
+    lastSeq number' . ($value !== null ? (' := ' . $value) : ''). ';
+begin' . ($value === null ? '
+    SELECT MAX("' . $tableSchema->getPrimaryKey()[0] . '") + 1 INTO lastSeq FROM "' . $tableSchema->getName() . '";' : '') . '
+    if lastSeq IS NULL then lastSeq := 1; end if;
+    execute immediate \'DROP SEQUENCE "' . $sequenceName . '"\';
+    execute immediate \'CREATE SEQUENCE "' . $sequenceName . '" START WITH \' || lastSeq || \' INCREMENT BY 1 NOMAXVALUE NOCACHE\';
+end;';
     }
 }
