+++ src/PDO/QueryBuilderPDOOracle.php
@@ -4,33 +4,13 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Oracle\PDO;
 
-use Generator;
-use Throwable;
-use Yiisoft\Db\Command\CommandInterface;
-use Yiisoft\Db\Exception\Exception;
-use Yiisoft\Db\Exception\InvalidArgumentException;
-use Yiisoft\Db\Exception\InvalidConfigException;
-use Yiisoft\Db\Exception\NotSupportedException;
-use Yiisoft\Db\Expression\ExpressionInterface;
-use Yiisoft\Db\Oracle\Builder\InConditionBuilder;
-use Yiisoft\Db\Oracle\Builder\LikeConditionBuilder;
 use Yiisoft\Db\Oracle\DDLQueryBuilder;
 use Yiisoft\Db\Oracle\DMLQueryBuilder;
-use Yiisoft\Db\Query\Conditions\InCondition;
-use Yiisoft\Db\Query\Conditions\LikeCondition;
-use Yiisoft\Db\Query\Query;
+use Yiisoft\Db\Oracle\DQLQueryBuilder;
 use Yiisoft\Db\Query\QueryBuilder;
-use Yiisoft\Db\Schema\ColumnSchemaBuilder;
 use Yiisoft\Db\Schema\QuoterInterface;
 use Yiisoft\Db\Schema\Schema;
 use Yiisoft\Db\Schema\SchemaInterface;
-use Yiisoft\Strings\NumericHelper;
-
-use function array_merge;
-use function implode;
-use function is_float;
-use function is_string;
-use function reset;
 
 /**
  * QueryBuilder is the query builder for Oracle databases.
@@ -65,320 +45,15 @@ final class QueryBuilderPDOOracle extends QueryBuilder
     ];
     private DDLQueryBuilder $ddlBuilder;
     private DMLQueryBuilder $dmlBuilder;
+    private DQLQueryBuilder $dqlBuilder;
 
     public function __construct(
-        private CommandInterface $command,
-        private QuoterInterface $quoter,
-        private SchemaInterface $schema
+        protected QuoterInterface $quoter,
+        protected SchemaInterface $schema
     ) {
         $this->ddlBuilder = new DDLQueryBuilder($this);
         $this->dmlBuilder = new DMLQueryBuilder($this);
-        parent::__construct($quoter, $schema, $this->ddlBuilder, $this->dmlBuilder);
-    }
-
-    public function addForeignKey(
-        string $name,
-        string $table,
-        $columns,
-        string $refTable,
-        $refColumns,
-        ?string $delete = null,
-        ?string $update = null
-    ): string {
-        $sql = 'ALTER TABLE ' . $this->quoter->quoteTableName($table)
-            . ' ADD CONSTRAINT ' . $this->quoter->quoteColumnName($name)
-            . ' FOREIGN KEY (' . $this->buildColumns($columns) . ')'
-            . ' REFERENCES ' . $this->quoter->quoteTableName($refTable)
-            . ' (' . $this->buildColumns($refColumns) . ')';
-
-        if ($delete !== null) {
-            $sql .= ' ON DELETE ' . $delete;
-        }
-
-        if ($update !== null) {
-            throw new Exception('Oracle does not support ON UPDATE clause.');
-        }
-
-        return $sql;
-    }
-
-    /**
-     * Builds a SQL statement for changing the definition of a column.
-     *
-     * @param string $table the table whose column is to be changed. The table name will be properly quoted by the
-     * method.
-     * @param string $column the name of the column to be changed. The name will be properly quoted by the method.
-     * @param ColumnSchemaBuilder|string $type the new column type. The [[getColumnType]] method will be invoked to convert abstract column
-     * type (if any) into the physical one. Anything that is not recognized as abstract type will be kept in the
-     * generated SQL.
-     *
-     * For example, 'string' will be turned into 'varchar(255)', while 'string not null' will become
-     * 'varchar(255) not null'.
-     *
-     * @return string the SQL statement for changing the definition of a column.
-     */
-    public function alterColumn(string $table, string $column, ColumnSchemaBuilder|string $type): string
-    {
-        $type = $this->getColumnType($type);
-
-        return 'ALTER TABLE '
-            . $this->quoter->quoteTableName($table)
-            . ' MODIFY '
-            . $this->quoter->quoteColumnName($column)
-            . ' ' . $this->getColumnType($type);
-    }
-
-    /**
-     * Generates a batch INSERT SQL statement.
-     *
-     * For example,
-     *
-     * ```php
-     * $sql = $queryBuilder->batchInsert('user', ['name', 'age'], [
-     *     ['Tom', 30],
-     *     ['Jane', 20],
-     *     ['Linda', 25],
-     * ]);
-     * ```
-     *
-     * Note that the values in each row must match the corresponding column names.
-     *
-     * @param string $table the table that new rows will be inserted into.
-     * @param array $columns the column names.
-     * @param Generator|iterable $rows the rows to be batched inserted into the table.
-     * @param array $params
-     *
-     * @throws \Exception|InvalidArgumentException
-     *
-     * @return string the batch INSERT SQL statement.
-     *
-     * @psalm-suppress MixedArrayOffset
-     */
-    public function batchInsert(string $table, array $columns, iterable|Generator $rows, array &$params = []): string
-    {
-        if (empty($rows)) {
-            return '';
-        }
-
-        $schema = $this->schema;
-
-        if (($tableSchema = $schema->getTableSchema($table)) !== null) {
-            $columnSchemas = $tableSchema->getColumns();
-        } else {
-            $columnSchemas = [];
-        }
-
-        $values = [];
-
-        /** @psalm-var string[][] $rows */
-        foreach ($rows as $row) {
-            $vs = [];
-            foreach ($row as $i => $value) {
-                if (isset($columns[$i], $columnSchemas[$columns[$i]])) {
-                    /** @var mixed $value */
-                    $value = $columnSchemas[$columns[$i]]->dbTypecast($value);
-                }
-
-                if (is_string($value)) {
-                    /** @var mixed $value */
-                    $value = $this->quoter->quoteValue($value);
-                } elseif (is_float($value)) {
-                    /* ensure type cast always has . as decimal separator in all locales */
-                    $value = NumericHelper::normalize($value);
-                } elseif ($value === false) {
-                    $value = 0;
-                } elseif ($value === null) {
-                    $value = 'NULL';
-                } elseif ($value instanceof ExpressionInterface) {
-                    $value = $this->buildExpression($value, $params);
-                }
-
-                /** @var mixed */
-                $vs[] = $value;
-            }
-
-            /** @psalm-var string[] $vs */
-            $values[] = '(' . implode(', ', $vs) . ')';
-        }
-
-        if (empty($values)) {
-            return '';
-        }
-
-        /** @psalm-var string[] $columns */
-        foreach ($columns as $i => $name) {
-            $columns[$i] = $this->quoter->quoteColumnName($name);
-        }
-
-        $tableAndColumns = ' INTO ' . $this->quoter->quoteTableName($table)
-            . ' (' . implode(', ', $columns) . ') VALUES ';
-
-        return 'INSERT ALL ' . $tableAndColumns . implode($tableAndColumns, $values) . ' SELECT 1 FROM SYS.DUAL';
-    }
-
-    public function buildOrderByAndLimit(string $sql, array $orderBy, $limit, $offset, array &$params = []): string
-    {
-        $orderBy = $this->buildOrderBy($orderBy, $params);
-
-        if ($orderBy !== '') {
-            $sql .= $this->separator . $orderBy;
-        }
-
-        $filters = [];
-
-        if ($this->hasOffset($offset)) {
-            $filters[] = 'rowNumId > ' . (string) $offset;
-        }
-
-        if ($this->hasLimit($limit)) {
-            $filters[] = 'rownum <= ' . (string) $limit;
-        }
-
-        if (empty($filters)) {
-            return $sql;
-        }
-
-        $filter = implode(' AND ', $filters);
-        return <<<SQL
-        WITH USER_SQL AS ($sql), PAGINATION AS (SELECT USER_SQL.*, rownum as rowNumId FROM USER_SQL)
-        SELECT * FROM PAGINATION WHERE $filter
-        SQL;
-    }
-
-    public function checkIntegrity(string $schema = '', string $table = '', bool $check = true): string
-    {
-        throw new NotSupportedException('Oracle does not support enabling/disabling integrity check.');
-    }
-
-    public function command(): CommandInterface
-    {
-        return $this->command;
-    }
-
-    /**
-     * Builds a SQL statement for dropping an index.
-     *
-     * @param string $name the name of the index to be dropped. The name will be properly quoted by the method.
-     * @param string $table the table whose index is to be dropped. The name will be properly quoted by the method.
-     *
-     * @return string the SQL statement for dropping an index.
-     */
-    public function dropIndex(string $name, string $table): string
-    {
-        return 'DROP INDEX ' . $this->quoter->quoteTableName($name);
-    }
-
-    /**
-     * Creates a SQL statement for resetting the sequence value of a table's primary key.
-     *
-     * The sequence will be reset such that the primary key of the next new row inserted will have the specified value
-     * or 1.
-     *
-     * @param string $tableName the name of the table whose primary key sequence will be reset.
-     * @param array|int|string|null $value the value for the primary key of the next new row inserted. If this is not
-     * set, the next new row's primary key will have a value 1.
-     *
-     * @throws Exception|InvalidArgumentException|InvalidConfigException|Throwable
-     */
-    public function executeResetSequence(string $tableName, array|int|string $value = null): void
-    {
-        $tableSchema = $this->schema->getTableSchema($tableName);
-
-        if ($tableSchema === null) {
-            throw new InvalidArgumentException("Unknown table: $tableName");
-        }
-
-        if ($tableSchema->getSequenceName() === null) {
-            throw new InvalidArgumentException("There is no sequence associated with table: $tableName");
-        }
-
-        if ($value !== null) {
-            $value = (int) $value;
-        } else {
-            if (count($tableSchema->getPrimaryKey()) > 1) {
-                throw new InvalidArgumentException(
-                    "Can't reset sequence for composite primary key in table: $tableName"
-                );
-            }
-            $value = (int) $this->command->setSql(
-                'SELECT MAX("' . $tableSchema->getPrimaryKey()[0] . '") FROM "' . $tableSchema->getName() . '"'
-            )->queryScalar() + 1;
-        }
-
-        /**
-         *  Oracle needs at least two queries to reset sequence (see adding transactions and/or use alter method to
-         *  avoid grants' issue?)
-         */
-        $this->command->setSQl('DROP SEQUENCE "' . (string) $tableSchema->getSequenceName() . '"')->execute();
-        $this->command->setSql(
-            'CREATE SEQUENCE "' .
-            (string) $tableSchema->getSequenceName() .
-            '" START WITH ' .
-            $value .
-            ' INCREMENT BY 1 NOMAXVALUE NOCACHE'
-        )->execute();
-    }
-
-    public function prepareInsertValues(string $table, $columns, array $params = []): array
-    {
-        /**
-         * @var array $names
-         * @var array $placeholders
-         */
-        [$names, $placeholders, $values, $params] = parent::prepareInsertValues($table, $columns, $params);
-
-        if (!$columns instanceof Query && empty($names)) {
-            $tableSchema = $this->schema->getTableSchema($table);
-
-            if ($tableSchema !== null) {
-                $tableColumns = $tableSchema->getColumns();
-                $columns = !empty($tableSchema->getPrimaryKey())
-                    ? $tableSchema->getPrimaryKey() : [reset($tableColumns)->getName()];
-                foreach ($columns as $name) {
-                    /** @var mixed */
-                    $names[] = $this->quoter->quoteColumnName($name);
-                    $placeholders[] = 'DEFAULT';
-                }
-            }
-        }
-
-        return [$names, $placeholders, $values, $params];
-    }
-
-    public function quoter(): QuoterInterface
-    {
-        return $this->quoter;
-    }
-
-    /**
-     * Builds a SQL statement for renaming a DB table.
-     *
-     * @param string $oldName
-     * @param string $newName the new table name. The name will be properly quoted by the method.
-     *
-     * @return string the SQL statement for renaming a DB table.
-     */
-    public function renameTable(string $oldName, string $newName): string
-    {
-        return 'ALTER TABLE ' . $this->quoter->quoteTableName($oldName) . ' RENAME TO ' .
-            $this->quoter->quoteTableName($newName);
-    }
-
-    public function selectExists(string $rawSql): string
-    {
-        return 'SELECT CASE WHEN EXISTS(' . $rawSql . ') THEN 1 ELSE 0 END FROM DUAL';
-    }
-
-    public function schema(): SchemaInterface
-    {
-        return $this->schema;
-    }
-
-    protected function defaultExpressionBuilders(): array
-    {
-        return array_merge(parent::defaultExpressionBuilders(), [
-            InCondition::class => InConditionBuilder::class,
-            LikeCondition::class => LikeConditionBuilder::class,
-        ]);
+        $this->dqlBuilder = new DQLQueryBuilder($this);
+        parent::__construct($quoter, $schema, $this->ddlBuilder, $this->dmlBuilder, $this->dqlBuilder);
     }
 }
